requirements.txt
torch
torchaudio
librosa
numpy
pandas
scikit-learn
hydra-core
mlflow
matplotlib
seaborn
tqdm
